moving function expressions [spark tests]
