+++ crates/sail-plan/src/extension/function/hash/mod.rs
@@ -1,0 +1,4 @@
+pub mod spark_crc32;
+pub mod spark_murmur3_hash;
+pub mod spark_sha1;
+pub mod spark_xxhash64;
