+++ crates/sail-plan/src/extension/function/hash/spark_sha1.rs
