+++ crates/sail-plan/src/extension/function/misc/bitmap_count.rs
