+++ crates/sail-plan/src/extension/function/misc/mod.rs
@@ -1,1 +1,4 @@
+pub mod bitmap_count;
+pub mod raise_error;
+pub mod spark_aes;
 pub mod version;
