+++ crates/sail-plan/src/extension/function/misc/raise_error.rs
