+++ crates/sail-plan/src/extension/function/mod.rs
@@ -1,5 +1,4 @@
 pub mod array;
-pub mod bitmap_count;
 pub mod collection;
 pub mod csv;
 pub mod datetime;
@@ -8,6 +7,7 @@ pub mod error_utils;
 pub mod explode;
 mod functions_nested_utils;
 mod functions_utils;
+pub mod hash;
 pub mod kurtosis;
 pub mod map;
 pub mod math;
@@ -15,15 +15,9 @@ pub mod max_min_by;
 pub mod misc;
 pub mod mode;
 pub mod multi_expr;
-pub mod raise_error;
 pub mod skewness;
-pub mod spark_aes;
-pub mod spark_crc32;
 mod spark_hash_utils;
-pub mod spark_murmur3_hash;
-pub mod spark_sha1;
 pub mod spark_to_string;
-pub mod spark_xxhash64;
 pub mod string;
 pub mod struct_function;
 pub mod table_input;
