+++ crates/sail-plan/src/function/scalar/misc.rs
@@ -8,12 +8,12 @@ use sail_catalog::utils::quote_namespace_if_needed;
 use sail_common_datafusion::extension::SessionExtensionAccessor;
 
 use crate::error::{PlanError, PlanResult};
-use crate::extension::function::bitmap_count::BitmapCount;
-use crate::extension::function::misc::version::SparkVersion;
-use crate::extension::function::raise_error::RaiseError;
-use crate::extension::function::spark_aes::{
+use crate::extension::function::misc::bitmap_count::BitmapCount;
+use crate::extension::function::misc::raise_error::RaiseError;
+use crate::extension::function::misc::spark_aes::{
     SparkAESDecrypt, SparkAESEncrypt, SparkTryAESDecrypt, SparkTryAESEncrypt,
 };
+use crate::extension::function::misc::version::SparkVersion;
 use crate::function::common::{ScalarFunction, ScalarFunctionInput};
 use crate::utils::ItemTaker;
